requirements.txt
argparse
docker-py==1.1.0